refactor binance vendor services
