+++ apps/vendor-binance/src/index.ts
@@ -1,4 +1,7 @@
-import './interest_rate';
-import './quote';
-import './legacy_index';
-import './product';
+import './account';
+import './public-data/interest_rate';
+import './order-actions-with-credential';
+import './order-actions';
+import './public-data/quote';
+import './transfer';
+import './public-data/product';
